servers/ai-training/yolov8/requirements.txt
ultralytics==8.1.0
opencv-python==4.8.1.78
pyyaml==6.0.1
numpy>=1.22.0
torch>=1.8.0
torchvision>=0.9.0
pillow>=7.1.2
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4
rtsp